app/agentic/eol/requirements.txt
fastapi==0.121.2  # Required by agent-framework-ag-ui prerelease
uvicorn[standard]==0.38.0
pydantic==2.12.4

# OPTIONAL: Agent Framework (for workflow orchestrator only)
# Only needed if USE_WORKFLOW_ORCHESTRATOR=true environment variable is set
# Can be omitted for production deployments using legacy orchestrator
agent-framework==1.0.0rc1  # Latest release (Feb 19, 2026)
agent-framework-ag-ui==1.0.0b260219
agent-framework-chatkit==1.0.0b260219
# agent-framework-core==1.0.0rc1 requires opentelemetry-api>=1.39.0 — let pip resolve
# Pin opentelemetry-semantic-conventions-ai to 0.4.13 — v0.4.14 removed LLM_SYSTEM which RC1 observability.py still uses
opentelemetry-semantic-conventions-ai==0.4.13
microsoft-agents-copilotstudio-client>=0.5.0,<0.7.0
microsoft-agents-hosting-core>=0.5.0,<0.7.0
microsoft-agents-activity>=0.5.0,<0.7.0
powerfx>=0.0.31,<0.1.0
clr-loader>=0.2.7,<0.3.0
azure-identity==1.17.1
azure-core==1.36.0
azure-monitor-query==1.4.0
# Azure Management SDK for workflow orchestrator and SRE operations
azure-mgmt-resource==23.1.1
azure-mgmt-compute==33.0.0
azure-mgmt-network==27.0.0
azure-mgmt-monitor==6.0.2
azure-mgmt-storage==21.2.1
azure-mgmt-resourcehealth==1.0.0b1
azure-mgmt-resourcegraph==8.0.1  # For dependency mapping in SRE MCP server
azure-mgmt-web==7.0.0  # For App Service operations in SRE MCP server
azure-mgmt-redis==14.0.0  # For Redis cache operations in SRE MCP server
azure-mgmt-containerservice==20.0.0  # For AKS operations in SRE MCP server
openai==2.21.0
tiktoken==0.12.0
a2a-sdk==0.3.12
tqdm==4.67.1
aiohttp==3.10.5
httpx==0.28.1
cachetools==5.5.0  # TTL cache for CVE data L1 caching
beautifulsoup4==4.12.3
soupsieve==2.5
lxml==5.3.0
pdfplumber==0.11.4
jinja2==3.1.4
python-multipart==0.0.9
pytest==8.3.2
pytest-asyncio==0.23.8
azure-cosmos==4.7.0
aiofiles==24.1.0
email-validator==2.1.0
PyYAML==6.0.2
playwright==1.50.0
# Azure AI Agent Service dependencies (Modern replacement for Bing Search)
azure-ai-projects==2.0.0b3
azure-ai-agents==1.2.0b5  # pinned by agent-framework-azure-ai==1.0.0rc1
azure-search-documents==11.5.3
html2text==2025.4.15
# Azure MCP Server integration
mcp==1.25.0
python-dotenv==1.2.1
# Microsoft Teams notifications for SRE agent
pymsteams==0.2.2
apscheduler>=3.10.4  # Scheduler for SRE periodic tasks (Phase 3)
sse-starlette==3.0.3
uvloop==0.22.1
websockets==15.0.1
watchfiles==1.1.1
httptools==0.7.1
# PDF generation for CVE dashboard reports
weasyprint==62.3